% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{check_tri_params}
\alias{check_tri_params}
\title{Check triangular distribution parameters}
\usage{
check_tri_params(min, max, mode)
}
\arguments{
\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
No return value. Stops with an error if parameters are invalid.
}
\description{
Check triangular distribution parameters
}
\keyword{internal}
