% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function of the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector of quantiles.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode of the distribution.}
}
\value{
A numeric vector with cumulative probabilities.
}
\description{
Computes the cumulative probability of a triangular distribution.
}
\examples{
ptriang(c(0, 0.5, 1), min = 0, max = 1, mode = 0.5)
}
