% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function of the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Numeric vector of probabilities between 0 and 1.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode of the distribution.}
}
\value{
A numeric vector with quantiles.
}
\description{
Computes the inverse cumulative probability of a triangular distribution.
}
\examples{
qtriang(c(0.25, 0.5, 0.75), min = 0, max = 1, mode = 0.5)
}
