% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{recycle_triang_args}
\alias{recycle_triang_args}
\title{Recycle vectors using base R recycling rules}
\usage{
recycle_triang_args(...)
}
\description{
Recycle vectors using base R recycling rules
}
\keyword{internal}
