% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{triang_dist}
\alias{triang_dist}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{Triangular Distribution}
\usage{
dtriang(x, min, max, mode)

ptriang(q, min, max, mode)

qtriang(p, min, max, mode)

rtriang(n, min, max, mode)
}
\arguments{
\item{x, q}{Quantile.}

\item{min}{Minimum value.}

\item{max}{Maximum value.}

\item{mode}{Mode value.
#'}

\item{p}{Probability.}

\item{n}{Observations to generate.}
}
\value{
\code{dtriang} Returns density value.

\code{ptriang} Returns cumulative density value.

\code{qtriang} Returns the quantile value.

Vector of length \code{n} with randomly generated values.
}
\description{
Density, distribution , quantile function and random generation for the
triangular distribution with parameters min, max and mode.
}
