% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{triangdist}
\alias{triangdist}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{Triangular Distribution}
\usage{
dtriang(x, min = 0, max = 1, mode = 0.5)

ptriang(q, min = 0, max = 1, mode = 0.5)

qtriang(p, min = 0, max = 1, mode = 0.5)

rtriang(n, min = 0, max = 1, mode = 0.5)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{peak of the distribution (c).}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
\code{dtriang} gives the density, \code{ptriang} gives the distribution function,
\code{qtriang} gives the quantile function, and \code{rtriang} generates random deviates.
}
\description{
Density, distribution function, quantile function and random
generation for the triangular distribution with parameters min, max and mode.
}
