% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{triangular}
\alias{triangular}
\alias{dtriang}
\alias{ptriang}
\alias{qtriang}
\alias{rtriang}
\title{Triangular Distribution}
\usage{
dtriang(x, min, max, mode)

ptriang(q, min, max, mode)

qtriang(p, min, max, mode)

rtriang(n, min, max, mode)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{min}{lower limit.}

\item{max}{upper limit.}

\item{mode}{mode of the distribution.}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\description{
Density, distribution function, quantile function and random generation
for the triangular distribution with parameters min, max and mode.
}
