% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{validate_triang_params}
\alias{validate_triang_params}
\title{Validate triangular distribution parameters}
\usage{
validate_triang_params(min, max, mode)
}
\description{
Validate triangular distribution parameters
}
\keyword{internal}
