% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{numeric vector}

\item{min}{left limit}

\item{max}{right limit}

\item{mode}{mode}
}
\value{
PDF result
}
\description{
Triangular density
}
