% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular Distribution Density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{A numeric vector of quantiles.}

\item{min}{The lower limit of the distribution.}

\item{max}{The upper limit of the distribution.}

\item{mode}{The mode (peak) of the distribution.}
}
\value{
A numeric vector containing the density values
}
\description{
Computes the probability density function (PDF) of triangular distribution.
}
\examples{
dtriang(0.5, 0, 1, 0.5)
dtriang(c(0.2, 0.5, 0.8), 0, 1, 0.5)
dtriang(-1, 0, 1, 0.5)  # fuera de rango, devuelve 0
}
