% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density of the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Vector of values.}

\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
Numeric vector of density values.
}
\description{
Density of the triangular distribution
}
