% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{numeric vector}

\item{min}{lower bound}

\item{max}{upper bound}

\item{mode}{mode}
}
\value{
density values
}
\description{
Triangular density
}
