% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density of the Triangular Distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{min}{Numeric vector of lower limits.}

\item{max}{Numeric vector of upper limits.}

\item{mode}{Numeric vector of modes.}
}
\value{
A numeric vector with density values.
}
\description{
Computes the density of the triangular distribution.
}
\examples{
dtriang(0.5, min = 0, max = 1, mode = 0.5)
}
