% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtriang.R
\name{dtriang}
\alias{dtriang}
\title{Density of the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector of values}

\item{min}{Lower bound (a)}

\item{max}{Upper bound (b)}

\item{mode}{Mode (c)}
}
\value{
Numeric vector of densities
}
\description{
Computes the probability density function of a triangular distribution.
}
