% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density of the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector of quantiles.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode of the distribution.}
}
\value{
A numeric vector with the density values.
}
\description{
Computes the density of a triangular distribution.
}
\examples{
dtriang(c(0, 0.5, 1), min = 0, max = 1, mode = 0.5)
}
