% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density function for the Triangular Distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{vector of quantiles.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{the mode of the distribution (c).}
}
\value{
A numeric vector representing the density values.
}
\description{
Computes the probability density function of the triangular distribution.
}
