% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density function for the triangular distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{vector of quantiles}

\item{min}{lower limit}

\item{max}{upper limit}

\item{mode}{mode}
}
\value{
A numeric vector of densities
}
\description{
Density function for the triangular distribution
}
