% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular distribution density function}
\usage{
dtriang(x, lower, upper, mode)
}
\arguments{
\item{x}{numeric vector of quantiles.}

\item{lower}{lower bound of the distribution.}

\item{upper}{upper bound of the distribution.}

\item{mode}{mode of the distribution, must satisfy lower <= mode <= upper.}
}
\value{
numeric vector with density values.
}
\description{
Computes the probability density function of the triangular distribution.
}
