% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular Distribution Density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{vector of quantiles.}

\item{min}{lower limit.}

\item{max}{upper limit.}

\item{mode}{the mode.}
}
\value{
A vector of densities.
}
\description{
Triangular Distribution Density
}
