% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular Distribution Density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{vector of quantiles.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
A numeric vector of densities.
}
\description{
Computes the probability density function of the triangular distribution.
}
\details{
The height of the triangle is calculated as 2 / (max - min).
}
