% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Density function for the Triangular Distribution}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Numeric vector of quantiles}

\item{min}{Lower limit (a)}

\item{max}{Upper limit (b)}

\item{mode}{Mode (c)}
}
\description{
Density function for the Triangular Distribution
}
