% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular density}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{Vector of values}

\item{min}{Lower limit}

\item{max}{Upper limit}

\item{mode}{Mode (peak)}
}
\value{
Density values
}
\description{
Triangular density
}
\examples{
dtriang(0.5, 0, 1, 0.5)
}
