% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{dtriang}
\alias{dtriang}
\title{Triangular Distribution PDF}
\usage{
dtriang(x, min, max, mode)
}
\arguments{
\item{x}{vector of quantiles.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{mode of the distribution (c).}
}
\value{
A vector of densities.
}
\description{
Density function (dtriang)
}
