% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{numeric vector}

\item{min}{left limit}

\item{max}{right limit}

\item{mode}{mode}
}
\value{
CDF result
}
\description{
Triangular distribution
}
