% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Cumulative Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{A numeric vector of quantiles.}

\item{min}{The lower limit of the distribution.}

\item{max}{The upper limit of the distribution.}

\item{mode}{The mode (peak) of the distribution.}
}
\value{
A numeric vector containing the cumulative probability.
}
\description{
Computes the cumulative probability (CDF) of the triangular distribution.
}
