% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function of the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Vector of values.}

\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
Numeric vector of cumulative probabilities.
}
\description{
Distribution function of the triangular distribution
}
