% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular distribution function}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{quantiles}

\item{min}{lower bound}

\item{max}{upper bound}

\item{mode}{mode}
}
\value{
probabilities
}
\description{
Triangular distribution function
}
