% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution Function of the Triangular Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector of quantiles.}

\item{min}{Numeric vector of lower limits.}

\item{max}{Numeric vector of upper limits.}

\item{mode}{Numeric vector of modes.}
}
\value{
A numeric vector with cumulative probabilities.
}
\description{
Computes the cumulative distribution function of the triangular distribution.
}
\examples{
ptriang(0.5, min = 0, max = 1, mode = 0.5)
}
