% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function for the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Vector of quantiles}

\item{min}{Lower limit of the distribution}

\item{max}{Upper limit of the distribution}

\item{mode}{Mode of the distribution}
}
\value{
A numeric vector of cumulative probabilities
}
\description{
Distribution function for the triangular distribution
}
