% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptriang.R
\name{ptriang}
\alias{ptriang}
\title{CDF of the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector of values}

\item{min}{Lower bound (a)}

\item{max}{Upper bound (b)}

\item{mode}{Mode (c)}
}
\value{
Numeric vector of probabilities
}
\description{
Computes the cumulative distribution function of a triangular distribution.
}
