% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function for the Triangular Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{vector of quantiles.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{the mode of the distribution (c).}
}
\value{
A numeric vector representing the cumulative probability values.
}
\description{
Computes the cumulative distribution function (CDF) of the triangular distribution.
}
