% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function for the triangular distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{vector of quantiles}

\item{min}{lower limit}

\item{max}{upper limit}

\item{mode}{mode}
}
\value{
A numeric vector of probabilities
}
\description{
Distribution function for the triangular distribution
}
