% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular distribution cumulative distribution function}
\usage{
ptriang(q, lower, upper, mode)
}
\arguments{
\item{q}{numeric vector of quantiles.}

\item{lower}{lower bound of the distribution.}

\item{upper}{upper bound of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
numeric vector with cumulative probabilities.
}
\description{
Computes the cumulative distribution function of the triangular distribution.
}
