% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular Cumulative Distribution Function}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{vector of quantiles.}

\item{min}{lower limit.}

\item{max}{upper limit.}

\item{mode}{the mode.}
}
\value{
A vector of probabilities.
}
\description{
Triangular Cumulative Distribution Function
}
