% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution Function of the Triangular Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Vector of quantiles.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of probabilities.
}
\description{
Calculates the cumulative distribution function (CDF)
for a triangular distribution.
}
