% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular Distribution Cumulative Distribution Function}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{vector of quantiles.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
A numeric vector of probabilities.
}
\description{
Computes the cumulative distribution function (CDF) of the triangular distribution.
}
