% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function for the Triangular Distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Numeric vector of quantiles}

\item{min}{Lower limit (a)}

\item{max}{Upper limit (b)}

\item{mode}{Mode (c)}
}
\description{
Distribution function for the Triangular Distribution
}
