% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Triangular cumulative distribution}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{Quantiles}

\item{min}{Lower limit}

\item{max}{Upper limit}

\item{mode}{Mode}
}
\value{
CDF values
}
\description{
Triangular cumulative distribution
}
\examples{
ptriang(0.5, 0, 1, 0.5)
}
