% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{ptriang}
\alias{ptriang}
\title{Distribution function (ptriang)}
\usage{
ptriang(q, min, max, mode)
}
\arguments{
\item{q}{vector of quantiles.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{mode of the distribution (c).}
}
\value{
A vector of probabilities.
}
\description{
Distribution function (ptriang)
}
