% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{numeric vector}

\item{min}{left limit}

\item{max}{right limit}

\item{mode}{mode}
}
\value{
Quantile result
}
\description{
Quantile function
}
