% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular Distribution Quantiles}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{A numeric vector of probabilities.}

\item{min}{The lower limit of the distribution.}

\item{max}{The upper limit of the distribution.}

\item{mode}{The mode (peak) of the distribution.}
}
\value{
A numeric vector containing the corresponding quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the triangular distribution.
}
\examples{
ptriang(0.5, 0, 1, 0.5)
ptriang(c(0.2, 0.5, 0.8), 0, 1, 0.5)
ptriang(0, 0, 1, 0.5)   # en el límite inferior, devuelve 0
ptriang(1, 0, 1, 0.5)   # en el límite superior, devuelve 1
}
