% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function of the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{min}{Lower limit.}

\item{max}{Upper limit.}

\item{mode}{Mode of the distribution.}
}
\value{
Numeric vector of quantiles.
}
\description{
Quantile function of the triangular distribution
}
