% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular quantile function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{probabilities}

\item{min}{lower bound}

\item{max}{upper bound}

\item{mode}{mode}
}
\value{
quantiles
}
\description{
Triangular quantile function
}
