% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile Function of the Triangular Distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Numeric vector of probabilities.}

\item{min}{Numeric vector of lower limits.}

\item{max}{Numeric vector of upper limits.}

\item{mode}{Numeric vector of modes.}
}
\value{
A numeric vector with quantiles.
}
\description{
Computes the quantile function of the triangular distribution.
}
\examples{
qtriang(0.5, min = 0, max = 1, mode = 0.5)
}
