% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtriang.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function of the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Numeric vector of probabilities}

\item{min}{Lower bound (a)}

\item{max}{Upper bound (b)}

\item{mode}{Mode (c)}
}
\value{
Numeric vector of quantiles
}
\description{
Computes the inverse CDF of a triangular distribution.
}
