% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the Triangular Distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{vector of probabilities.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{the mode of the distribution (c).}
}
\value{
A numeric vector representing the quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the triangular distribution.
}
