% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{vector of probabilities}

\item{min}{lower limit}

\item{max}{upper limit}

\item{mode}{mode}
}
\value{
A numeric vector of quantiles
}
\description{
Quantile function for the triangular distribution
}
