% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular distribution quantile function}
\usage{
qtriang(p, lower, upper, mode)
}
\arguments{
\item{p}{numeric vector of probabilities in interval 0, 1".}

\item{lower}{lower bound of the distribution.}

\item{upper}{upper bound of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
numeric vector of quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the triangular distribution.
}
