% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular Quantile Function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{vector of probabilities.}

\item{min}{lower limit.}

\item{max}{upper limit.}

\item{mode}{the mode.}
}
\value{
A vector of quantiles.
}
\description{
Triangular Quantile Function
}
