% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the triangular distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Vector of probabilities.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of quantiles.
}
\description{
Quantile function for the triangular distribution
}
