% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular Distribution Quantile Function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{vector of probabilities.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
A numeric vector of quantiles.
}
\description{
Computes the quantile function (inverse CDF) of the triangular distribution.
}
