% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function of the triangular distribution}
\usage{
qtriang(p, min = 0, max = 1, mode = 0.5)
}
\arguments{
\item{p}{Vector of probabilities. Must be in \verb{[0, 1]}.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode (peak) of the distribution.}
}
\value{
A numeric vector of quantiles.
}
\description{
Quantile function of the triangular distribution
}
\examples{
qtriang(0.5, min = 0, max = 1, mode = 0.5)
qtriang(c(0, 0.25, 0.5, 0.75, 1), min = 0, max = 1, mode = 0.3)
}
