% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function for the Triangular Distribution}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Numeric vector of probability}

\item{min}{Lower limit (a)}

\item{max}{Upper limit (b)}

\item{mode}{Mode (c)}
}
\description{
Quantile function for the Triangular Distribution
}
