% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Triangular quantile function}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{Probabilities}

\item{min}{Lower limit}

\item{max}{Upper limit}

\item{mode}{Mode}
}
\value{
Quantiles
}
\description{
Triangular quantile function
}
\examples{
qtriang(0.5, 0, 1, 0.5)
}
