% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{qtriang}
\alias{qtriang}
\title{Quantile function (qtriang)}
\usage{
qtriang(p, min, max, mode)
}
\arguments{
\item{p}{vector of probabilities.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{mode of the distribution (c).}
}
\value{
A vector of quantiles.
}
\description{
Quantile function (qtriang)
}
