% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation.}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{natural number. Number of random generations.}

\item{min}{left limit}

\item{max}{right limit}

\item{mode}{mode}
}
\value{
Random generation of length n
}
\description{
Random generation.
}
