% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Triangular Distribution Random Generation}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations.}

\item{min}{The lower limit of the distribution.}

\item{max}{The upper limit of the distribution.}

\item{mode}{The mode (peak) of the distribution.}
}
\value{
A numeric vector of random generated values.
}
\description{
Computes random deviates of the triangular distribution.
}
