% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{sample size}

\item{min}{lower bound}

\item{max}{upper bound}

\item{mode}{mode}
}
\value{
random sample
}
\description{
Random generation
}
