% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random Generation from the Triangular Distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations.}

\item{min}{Numeric vector of lower limits.}

\item{max}{Numeric vector of upper limits.}

\item{mode}{Numeric vector of modes.}
}
\value{
A numeric vector of random values.
}
\description{
Generates random values from the triangular distribution using inverse
transform sampling.
}
\examples{
rtriang(5, min = 0, max = 1, mode = 0.5)
}
