% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtriang.R
\name{rtriang}
\alias{rtriang}
\title{Random generation from triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations}

\item{min}{Lower bound (a)}

\item{max}{Upper bound (b)}

\item{mode}{Mode (c)}
}
\value{
Numeric vector of random values
}
\description{
Generates random values following a triangular distribution.
}
