% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode of the distribution.}
}
\value{
A numeric vector with random values.
}
\description{
Generates random values from a triangular distribution.
}
\examples{
set.seed(1)
rtriang(5, min = 0, max = 1, mode = 0.5)
}
