% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the Triangular Distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{the mode of the distribution (c).}
}
\value{
A numeric vector of random values.
}
\description{
Generates random deviates from the triangular distribution.
}
