% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{number of observations}

\item{min}{lower limit}

\item{max}{upper limit}

\item{mode}{mode}
}
\value{
A numeric vector of random values
}
\description{
Random generation for the triangular distribution
}
