% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation from triangular distribution}
\usage{
rtriang(n, lower, upper, mode)
}
\arguments{
\item{n}{number of observations to generate.}

\item{lower}{lower bound of the distribution.}

\item{upper}{upper bound of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
numeric vector of random samples.
}
\description{
Generates random samples from a triangular distrib using inverse transform.
}
