% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Triangular Random Generation}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{number of observations.}

\item{min}{lower limit.}

\item{max}{upper limit.}

\item{mode}{the mode.}
}
\value{
A vector of random numbers.
}
\description{
Triangular Random Generation
}
