% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random Generation for the Triangular Distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations. If length(n) > 1, the length is taken
to be the number required.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of random deviates.
}
\description{
Generates random deviates for a triangular distribution
using the inverse transform sampling method.
}
