% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of observations.}

\item{min}{Lower limit (a).}

\item{max}{Upper limit (b).}

\item{mode}{Mode (c).}
}
\value{
A numeric vector of random deviates.
}
\description{
Random generation for the triangular distribution
}
