% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Triangular Distribution Random Generation}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{min}{lower limit of the distribution.}

\item{max}{upper limit of the distribution.}

\item{mode}{mode of the distribution.}
}
\value{
A numeric vector of random deviates.
}
\description{
Generates random deviates from the triangular distribution.
}
