% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation for the triangular distribution}
\usage{
rtriang(n, min = 0, max = 1, mode = 0.5)
}
\arguments{
\item{n}{Number of random values to generate.}

\item{min}{Lower limit of the distribution.}

\item{max}{Upper limit of the distribution.}

\item{mode}{Mode (peak) of the distribution.}
}
\value{
A numeric vector of n random values.
}
\description{
Random generation for the triangular distribution
}
\examples{
set.seed(42)
rtriang(10, min = 0, max = 1, mode = 0.5)
}
