% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation from triangular distribution}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{Number of samples}

\item{min}{Lower limit}

\item{max}{Upper limit}

\item{mode}{Mode}
}
\value{
Random values
}
\description{
Random generation from triangular distribution
}
\examples{
rtriang(10, 0, 1, 0.5)
}
