% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_dist.R
\name{rtriang}
\alias{rtriang}
\title{Random generation (rtriang)}
\usage{
rtriang(n, min, max, mode)
}
\arguments{
\item{n}{number of observations.}

\item{min}{lower limit of the distribution (a).}

\item{max}{upper limit of the distribution (b).}

\item{mode}{mode of the distribution (c).}
}
\value{
A vector of random deviates.
}
\description{
Random generation (rtriang)
}
